requirements.txt
fastapi>=0.103.1,<0.110.0
uvicorn>=0.23.2,<0.30.0
sentence-transformers>=2.2.2,<3.0.0
faiss-cpu
pandas>=2.2.0  # Newer version for Python 3.13 compatibility
numpy>=1.25.2
scikit-learn>=1.3.0